more configs

+++ Assets/Scripts/Meta/Entities/Wheels.cs
@@ -3,6 +3,6 @@ namespace Meta.Entities
     public class Wheels
     {
         public string Id;
-        public int Price;
+        public long Price;
     }
 }
